Fix double invokation of callback if callback returns an error
This is partially based on the report and patch from
Jason Volk <jason@zemos.net>, however it actually fixes the issue
properly in both cases and doesn't rethrow the callbacks own exception
for error handling... It also improves the error message for aborted
callbacks by hardcoding the curl error code. This regularly happens when
you close Nheko and it cancels the sync request.